Move event bus pusher to its own common module
