+++ liveactivities/src/main/scala/com/gu/liveactivities/Lambda.scala
@@ -1,8 +1,8 @@
 package com.gu.liveactivities
 
-import aws.LiveActivityPusher
 import com.gu.liveactivities.service.{Authentication, LiveActivityChannelRepository}
 import com.gu.liveactivities.util.{Configuration, IosConfiguration, Logging}
+import com.gu.mobile.liveactivities.event.bus.LiveActivityPusher
 import software.amazon.awssdk.auth.credentials.{AwsCredentialsProviderChain, DefaultCredentialsProvider, ProfileCredentialsProvider}
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration
 import software.amazon.awssdk.regions.Region
